Add pixel shape option
